requirements.txt
loguru==0.5.3
mongo==0.2.0
pymongo==4.1.1
pyTelegramBotAPI==4.10.0
python-decouple==3.6
python-dotenv==0.20.0
python-utils==3.2.2
schedule==1.1.0
urllib3==1.26.5
requests==2.25.1
dnspython==2.2.1
git+https://github.com/JLeopolt/pytube.git
Flask==2.1.2
tenacity==6.3.1
aiohttp==3.8.3
humanize==4.4.0
Pillow==9.4.0
gunicorn==20.1.0
chess==1.9.4
CairoSVG==2.6.0